requirements.txt
dotmap
PyMcubes
pytorch-lightning==1.5.5
hydra-core==1.3.1
omegaconf==2.3.0
scikit-image==0.18.1
wandb
rich==12.5.1
h5py
pandas
open3d
trimesh
opencv-python
transformers
